refactor invoice Controller
