+++ controllers/erpControllers/invoiceController/create.js
@@ -1,0 +1,92 @@
+const mongoose = require('mongoose');
+
+const Model = mongoose.model('Invoice');
+
+const custom = require('@/controllers/middlewaresControllers/pdfController');
+
+const { calculate } = require('@/helpers');
+const schema = require('./schemaValidate');
+
+const create = async (req, res) => {
+  try {
+    let body = req.body;
+
+    const { error, value } = schema.validate(body);
+    if (error) {
+      const { details } = error;
+      return res.status(400).json({
+        success: false,
+        result: null,
+        message: details[0]?.message,
+      });
+    }
+
+    const { items = [], taxRate = 0, discount = 0 } = value;
+
+    // default
+    let subTotal = 0;
+    let taxTotal = 0;
+    let total = 0;
+
+    //Calculate the items array with subTotal, total, taxTotal
+    items.map((item) => {
+      let total = calculate.multiply(item['quantity'], item['price']);
+      //sub total
+      subTotal = calculate.add(subTotal, total);
+      //item total
+      item['total'] = total;
+    });
+    taxTotal = calculate.multiply(subTotal, taxRate);
+    total = calculate.add(subTotal, taxTotal);
+
+    body['subTotal'] = subTotal;
+    body['taxTotal'] = taxTotal;
+    body['total'] = total;
+    body['items'] = items;
+
+    let paymentStatus = calculate.sub(total, discount) === 0 ? 'paid' : 'unpaid';
+
+    body['paymentStatus'] = paymentStatus;
+    // Creating a new document in the collection
+    const result = await new Model(body).save();
+    const fileId = 'invoice-' + result._id + '.pdf';
+    const updateResult = await Model.findOneAndUpdate(
+      { _id: result._id },
+      { pdfPath: fileId },
+      {
+        new: true,
+      }
+    ).exec();
+    // Returning successfull response
+
+    custom.generatePdf('Invoice', { filename: 'invoice', format: 'A4' }, result);
+
+    // Returning successfull response
+    return res.status(200).json({
+      success: true,
+      result: updateResult,
+      message: 'Invoice created successfully',
+    });
+  } catch (err) {
+    console.log(err);
+    // If err is thrown by Mongoose due to required validations
+    if (err.name == 'ValidationError') {
+      return res.status(400).json({
+        success: false,
+        result: null,
+        error: err,
+        message: 'Required fields are not supplied',
+      });
+    } else {
+      // Server Error
+      return res.status(500).json({
+        success: false,
+        result: null,
+        error: err,
+        message: 'Oops there is an Error',
+      });
+    }
+  }
+};
+
+module.exports = create;
